ecommerce-tests/performance/requirements.txt
locust==2.16.0
python-dotenv==1.0.0
faker==19.1.0
matplotlib==3.7.2
pandas==2.0.3
